api fixed for uap and linux
