+++ src/System.Diagnostics.Process/tests/ProcessStartInfoTests.cs
@@ -988,46 +988,6 @@ namespace System.Diagnostics.Tests
             }
         }
 
-        [ConditionalTheory(typeof(PlatformDetection), nameof(PlatformDetection.IsNotWindowsNanoServer))] // No Notepad on Nano
-        [MemberData(nameof(UseShellExecute))]
-        [OuterLoop("Launches notepad")]
-        [PlatformSpecific(TestPlatforms.Windows)]
-        [SkipOnTargetFramework(TargetFrameworkMonikers.Uap, "WaitForInputIdle, ProcessName, and MainWindowTitle are not supported on UAP")]
-        public void StartInfo_NotepadWithContent_withArgumentList(bool useShellExecute)
-        {
-            string tempFile = GetTestFilePath() + ".txt";
-            File.WriteAllText(tempFile, $"StartInfo_NotepadWithContent({useShellExecute})");
-
-            ProcessStartInfo info = new ProcessStartInfo
-            {
-                UseShellExecute = useShellExecute,
-                FileName = @"notepad.exe",
-                Arguments = null,
-                WindowStyle = ProcessWindowStyle.Minimized
-            };
-
-            info.ArgumentList.Add(tempFile);
-
-            using (var process = Process.Start(info))
-            {
-                Assert.True(process != null, $"Could not start {info.FileName} {info.Arguments} UseShellExecute={info.UseShellExecute}");
-
-                try
-                {
-                    process.WaitForInputIdle(); // Give the file a chance to load
-                    Assert.Equal("notepad", process.ProcessName);
-
-                    // On some Windows versions, the file extension is not included in the title
-                    Assert.StartsWith(Path.GetFileNameWithoutExtension(tempFile), process.MainWindowTitle);
-                }
-                finally
-                {
-                    if (process != null && !process.HasExited)
-                        process.Kill();
-                }
-            }
-        }
-
         [ConditionalFact(typeof(PlatformDetection), nameof(PlatformDetection.IsNotWindowsNanoServer), // Nano does not support UseShellExecute
                                                     nameof(PlatformDetection.IsNotWindows8x))] // https://github.com/dotnet/corefx/issues/20388
         [OuterLoop("Launches notepad")]
